+++ mike/commands.py
@@ -26,17 +26,19 @@ def _add_redirect_to_commit(commit, template, src, dst):
         commit.add_file(git_utils.FileInfo(src, template.render(href=href)))
 
 
-def list_versions(branch='gh-pages'):
+def list_versions(branch='gh-pages', prefix=''):
     try:
         return Versions.loads(git_utils.read_file(
-            branch, versions_file, universal_newlines=True
+            branch, os.path.join(prefix, versions_file),
+            universal_newlines=True
         ))
     except git_utils.GitError:
         return Versions()
 
 
-def versions_to_file_info(versions):
-    return git_utils.FileInfo(versions_file, versions.dumps())
+def versions_to_file_info(versions, prefix=''):
+    return git_utils.FileInfo(os.path.join(prefix, versions_file),
+                              versions.dumps())
 
 
 def make_nojekyll():
@@ -44,21 +46,25 @@ def make_nojekyll():
 
 
 def deploy(site_dir, version, title=None, aliases=[], update_aliases=False,
-           redirect=True, template=None, *, branch='gh-pages', message=None):
+           redirect=True, template=None, *, branch='gh-pages', message=None,
+           prefix=''):
     if message is None:
         message = (
-            'Deployed {rev} to {doc_version} with MkDocs {mkdocs_version} ' +
-            'and mike {mike_version}'
+            'Deployed {rev} to {doc_version}{prefix} with MkDocs ' +
+            '{mkdocs_version} and mike {mike_version}'
         ).format(
             rev=git_utils.get_latest_commit('HEAD', short=True),
             doc_version=version,
+            prefix=' in {}'.format(prefix) if prefix else '',
             mkdocs_version=mkdocs_utils.version(),
             mike_version=app_version
         )
 
-    all_versions = list_versions(branch)
+    all_versions = list_versions(branch, prefix)
     info = all_versions.add(version, title, aliases, update_aliases)
     version_str = str(info.version)
+    destdir = os.path.join(prefix, version_str)
+    alias_destdirs = [os.path.join(prefix, i) for i in info.aliases]
 
     if redirect and info.aliases:
         t = _redirect_template(template)
@@ -67,9 +73,9 @@ def deploy(site_dir, version, title=None, aliases=[], update_aliases=False,
         commit.delete_files([version_str] + list(info.aliases))
 
         for f in git_utils.walk_real_files(site_dir):
-            canonical_file = f.copy(version_str, site_dir)
+            canonical_file = f.copy(destdir, site_dir)
             commit.add_file(canonical_file)
-            for d in info.aliases:
+            for d in alias_destdirs:
                 alias_file = f.copy(d, site_dir)
                 if redirect:
                     _add_redirect_to_commit(commit, t, alias_file.path,
@@ -77,27 +83,32 @@ def deploy(site_dir, version, title=None, aliases=[], update_aliases=False,
                 else:
                     commit.add_file(alias_file)
 
-        commit.add_file(versions_to_file_info(all_versions))
+        commit.add_file(versions_to_file_info(all_versions, prefix))
         commit.add_file(make_nojekyll())
 
 
-def delete(versions=None, all=False, *, branch='gh-pages', message=None):
+def delete(versions=None, all=False, *, branch='gh-pages', message=None,
+           prefix=''):
     if not all and versions is None:
         raise ValueError('specify `version` or `all`')
 
     if message is None:
         message = (
-            'Removed {doc_version} with mike {mike_version}'
+            'Removed {doc_version}{prefix} with mike {mike_version}'
         ).format(
             doc_version='everything' if all else ', '.join(versions),
+            prefix=' in {}'.format(prefix) if prefix else '',
             mike_version=app_version
         )
 
     with git_utils.Commit(branch, message) as commit:
         if all:
-            commit.delete_files('*')
+            if prefix:
+                commit.delete_files([prefix])
+            else:
+                commit.delete_files('*')
         else:
-            all_versions = list_versions(branch)
+            all_versions = list_versions(branch, prefix)
             try:
                 removed = all_versions.difference_update(versions)
             except KeyError as e:
@@ -105,15 +116,18 @@ def delete(versions=None, all=False, *, branch='gh-pages', message=None):
 
             for i in removed:
                 if isinstance(i, str):
-                    commit.delete_files([i])
+                    commit.delete_files([os.path.join(prefix, i)])
                 else:
-                    commit.delete_files([str(i.version)] + list(i.aliases))
-            commit.add_file(versions_to_file_info(all_versions))
+                    commit.delete_files(
+                        [os.path.join(prefix, str(i.version))] +
+                        [os.path.join(prefix, j) for j in i.aliases]
+                    )
+            commit.add_file(versions_to_file_info(all_versions, prefix))
 
 
 def alias(version, aliases, redirect=True, template=None, *, branch='gh-pages',
-          message=None):
-    all_versions = list_versions(branch)
+          message=None, prefix=''):
+    all_versions = list_versions(branch, prefix)
     try:
         real_version = all_versions.find(version, strict=True)[0]
     except KeyError as e:
@@ -121,61 +135,78 @@ def alias(version, aliases, redirect=True, template=None, *, branch='gh-pages',
 
     if message is None:
         message = (
-            'Copied {doc_version} to {aliases} with mike {mike_version}'
+            'Copied {doc_version} to {aliases}{prefix} with mike ' +
+            '{mike_version}'
         ).format(
             doc_version=real_version,
             aliases=', '.join(aliases),
+            prefix=' in {}'.format(prefix) if prefix else '',
             mike_version=app_version
         )
 
-    destdirs = all_versions.update(real_version, aliases=aliases)
+    new_aliases = all_versions.update(real_version, aliases=aliases)
+    destdirs = [os.path.join(prefix, i) for i in new_aliases]
+
     if redirect and destdirs:
         t = _redirect_template(template)
 
     with git_utils.Commit(branch, message) as commit:
         commit.delete_files(destdirs)
 
-        for canonical_file in git_utils.walk_files(branch, str(real_version)):
+        canonical_dir = os.path.join(prefix, str(real_version))
+        for canonical_file in git_utils.walk_files(branch, canonical_dir):
             for d in destdirs:
-                alias_file = canonical_file.copy(d, str(real_version))
+                alias_file = canonical_file.copy(d, canonical_dir)
                 if redirect:
                     _add_redirect_to_commit(commit, t, alias_file.path,
                                             canonical_file.path)
                 else:
                     commit.add_file(alias_file)
-        commit.add_file(versions_to_file_info(all_versions))
+        commit.add_file(versions_to_file_info(all_versions, prefix))
 
 
-def retitle(version, title, *, branch='gh-pages', message=None):
+def retitle(version, title, *, branch='gh-pages', message=None, prefix=''):
     if message is None:
         message = (
-            'Set title of {doc_version} to {title} with mike {mike_version}'
-        ).format(doc_version=version, title=title, mike_version=app_version)
+            'Set title of {doc_version} to {title}{prefix} with mike ' +
+            '{mike_version}'
+        ).format(
+            doc_version=version,
+            title=title,
+            prefix=' in {}'.format(prefix) if prefix else '',
+            mike_version=app_version
+        )
 
-    all_versions = list_versions(branch)
+    all_versions = list_versions(branch, prefix)
     try:
         all_versions.update(version, title)
     except KeyError:
         raise ValueError('version {} does not exist'.format(version))
 
     with git_utils.Commit(branch, message) as commit:
-        commit.add_file(versions_to_file_info(all_versions))
+        commit.add_file(versions_to_file_info(all_versions, prefix))
 
 
-def set_default(version, template=None, *, branch='gh-pages', message=None):
+def set_default(version, template=None, *, branch='gh-pages', message=None,
+                prefix=''):
     if message is None:
         message = (
-            'Set default version to {doc_version} with mike {mike_version}'
-        ).format(doc_version=version, mike_version=app_version)
+            'Set default version to {doc_version}{prefix} with mike ' +
+            '{mike_version}'
+        ).format(
+            doc_version=version,
+            prefix=' in {}'.format(prefix) if prefix else '',
+            mike_version=app_version
+        )
 
-    all_versions = list_versions(branch)
+    all_versions = list_versions(branch, prefix)
     if not all_versions.find(version):
         raise ValueError('version {} does not exist'.format(version))
 
     t = _redirect_template(template)
     with git_utils.Commit(branch, message) as commit:
         commit.add_file(git_utils.FileInfo(
-            'index.html', t.render(href=version)
+            os.path.join(prefix, 'index.html'), t.render(href=version)
         ))