+++ test/integration/test_retitle.py
@@ -7,25 +7,29 @@ from mike import git_utils, versions
 
 
 class RetitleTestCase(unittest.TestCase):
-    def _deploy(self, branch=None, versions=['1.0']):
-        branch_args = ['-b', branch] if branch else []
+    def _deploy(self, branch=None, versions=['1.0'], prefix=''):
+        extra_args = ['-b', branch] if branch else []
+        if prefix:
+            extra_args.extend(['--prefix', prefix])
         for i in versions:
-            assertPopen(['mike', 'deploy', i] + branch_args)
+            assertPopen(['mike', 'deploy', i] + extra_args)
 
-    def _test_retitle(self, expected_message=None):
+    def _test_retitle(self, expected_message=None, directory='.'):
         message = assertPopen(['git', 'log', '-1', '--pretty=%B']).rstrip()
         if expected_message:
             self.assertEqual(message, expected_message)
         else:
-            self.assertRegex(message,
-                             r'^Set title of \S+ to 1\.0\.1 with mike \S+$')
+            self.assertRegex(
+                message,
+                r'^Set title of \S+ to 1\.0\.1( in .*)? with mike \S+$'
+            )
 
-        assertDirectory('.', {
+        assertDirectory(directory, {
             'versions.json',
             '1.0/index.html',
         }, allow_extra=True)
 
-        with open('versions.json') as f:
+        with open(os.path.join(directory, 'versions.json')) as f:
             self.assertEqual(list(versions.Versions.loads(f.read())), [
                 versions.VersionInfo('1.0', '1.0.1'),
             ])
@@ -76,6 +80,12 @@ class TestRetitle(RetitleTestCase):
                      'commit message'])
         check_call_silent(['git', 'checkout', 'gh-pages'])
         self._test_retitle('commit message')
+
+    def test_prefix(self):
+        self._deploy(prefix='prefix')
+        assertPopen(['mike', 'retitle', '1.0', '1.0.1', '--prefix', 'prefix'])
+        check_call_silent(['git', 'checkout', 'gh-pages'])
+        self._test_retitle(directory='prefix')
 
     def test_push(self):
         self._deploy()